incapsulate pool in proto_bus

+++ buffer.h
@@ -18,7 +18,7 @@ public:
     static constexpr size_t kInvalidBuffer = std::numeric_limits<size_t>::max();
 
 public:
-    BufferPool(int start_size)
+    BufferPool(size_t start_size)
         : start_size_(start_size)
     {
     }
